2021/20/src/main_test.go
// Copyright 2021 Google LLC
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package main

import (
	"os"
	"strconv"
	"testing"

	"github.com/simmonmt/aoc/2021/common/logger"
	"github.com/simmonmt/aoc/2021/common/pos"
)

func TestCalcPosNum(t *testing.T) {
	type TestCase struct {
		in   []string
		want int
	}

	testCases := []TestCase{
		TestCase{
			in: []string{
				"...",
				"#..",
				".#.",
			},
			want: 34,
		},
	}

	for i, tc := range testCases {
		t.Run(strconv.Itoa(i), func(t *testing.T) {
			b := parseBoard(tc.in)
			b.Dump()
			if got := calcPosNum(b, pos.P2{1, 1}); got != tc.want {
				t.Errorf("calcPosNum = %v, want %v", got, tc.want)
			}
		})
	}
}

func TestMain(m *testing.M) {
	logger.Init(true)
	os.Exit(m.Run())
}
